Use fixed width integer types for cross-platform compatibility

+++ include/packets/packet_factory.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cstdint>
+
 class Packet;
 
 class PacketFactory
 {
 public:
-    static Packet* Create(int packetId);
+    static Packet* Create(uint8_t packetId);
 };
